Extract reporting library into org.osate.analysis.reporting bundle
Move the generic tabular-reporting code (Report/Section/Line/
ReportedCell/ReportSeverity, Csv/Excel/GenericExport, ReportUtils) out
of org.osate.analysis.flows into a new org.osate.analysis.reporting
bundle, renaming the packages org.osate.analysis.flows.reporting.* ->
org.osate.analysis.reporting.*. jxl.jar moves with the exporters; the
new bundle re-exports the jxl packages so flows still resolves them
transitively.

This severs the resource.budgets -> flows dependency: resource.budgets
only ever used the reporting code, never flow analysis. Also remove the
stale, unused org.osate.analysis.architecture requirement from
resource.budgets.

Register the new bundle in the analyses reactor, the OSATE plug-ins
feature, the javadoc dependency list (org.osate.help), and the JaCoCo
coverage report.
