+++ alisa/org.osate.verify/META-INF/MANIFEST.MF
@@ -25,7 +25,7 @@ Require-Bundle: org.eclipse.xtext;bundle-version="[2.20.0,3.0.0)";visibility:=re
  org.junit;bundle-version="[4.11.0,5.0.0)",
  org.osate.analysis.architecture;bundle-version="[3.1.0,4.0.0)",
  org.osate.analysis.flows;bundle-version="[6.0.0,7.0.0)",
- org.osate.analysis.resource.budgets;bundle-version="[4.3.0,5.0.0)",
+ org.osate.analysis.resource.budgets;bundle-version="[5.0.0,6.0.0)",
  org.osate.analysis.resource.management;bundle-version="[4.2.0,5.0.0)",
  org.eclipse.core.runtime;bundle-version="[3.10.0,4.0.0)",
  org.eclipse.core.resources;bundle-version="[3.9.1,4.0.0)",
