+++ analyses/org.osate.analysis.flows/src/org/osate/analysis/flows/model/LatencyContributor.java
@@ -35,8 +35,8 @@ import org.osate.aadl2.NamedElement;
 import org.osate.aadl2.instance.ConnectionInstance;
 import org.osate.aadl2.instance.InstanceObject;
 import org.osate.analysis.flows.internal.utils.FlowLatencyUtil;
-import org.osate.analysis.flows.reporting.model.Line;
-import org.osate.analysis.flows.reporting.model.ReportSeverity;
+import org.osate.analysis.reporting.model.Line;
+import org.osate.analysis.reporting.model.ReportSeverity;
 import org.osate.result.Diagnostic;
 import org.osate.result.DiagnosticType;
 import org.osate.result.Result;
