+++ analyses/org.osate.analysis.reporting/.gitignore
@@ -1,0 +1,2 @@
+/bin
+/target
