+++ analyses/org.osate.analysis.reporting/META-INF/MANIFEST.MF
@@ -1,0 +1,24 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %pluginName
+Bundle-SymbolicName: org.osate.analysis.reporting;singleton:=true
+Bundle-Version: 1.0.0.qualifier
+Bundle-Vendor: CMU/SEI
+Bundle-Localization: plugin
+Bundle-ClassPath: ., lib/jxl.jar
+Export-Package: org.osate.analysis.reporting.exporters,
+ org.osate.analysis.reporting.model,
+ jxl,
+ jxl.format,
+ jxl.write,
+ jxl.write.biff
+Require-Bundle: org.eclipse.core.runtime;bundle-version="[3.7.0,4.0.0)",
+ org.eclipse.core.resources;bundle-version="[3.7.0,4.0.0)",
+ org.eclipse.emf.ecore;bundle-version="[2.36.0,3.0.0)",
+ org.eclipse.emf.common;bundle-version="[2.11.1,3.0.0)",
+ org.osate.aadl2;bundle-version="[6.1.0,7.0.0)",
+ org.osate.aadl2.modelsupport;bundle-version="[8.1.0,9.0.0)",
+ org.osate.results;bundle-version="[3.2.0,4.0.0)"
+Bundle-ActivationPolicy: lazy
+Automatic-Module-Name: org.osate.analysis.reporting
+Bundle-RequiredExecutionEnvironment: JavaSE-21
