+++ analyses/org.osate.analysis.reporting/build.properties
@@ -1,7 +1,7 @@
-bin.includes = plugin.xml,\
-               plugin.properties,\
+bin.includes = plugin.properties,\
                META-INF/,\
-               .
+               .,\
+               lib/
 jars.compile.order = .
 source.. = src/
 output.. = bin/
