+++ analyses/org.osate.analysis.reporting/src/org/osate/analysis/reporting/model/Section.java
@@ -21,7 +21,7 @@
  * aries to this license with respect to the terms applicable to their Third Party Software. Third Party Software li-
  * censes only apply to the Third Party Software and not any other portion of this program or this program as a whole.
  */
-package org.osate.analysis.flows.reporting.model;
+package org.osate.analysis.reporting.model;
 
 import java.util.ArrayList;
 import java.util.List;
