+++ analyses/org.osate.analysis.resource.budgets/META-INF/MANIFEST.MF
@@ -2,7 +2,7 @@ Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: Resource Analysis Plug-in
 Bundle-SymbolicName: org.osate.analysis.resource.budgets;singleton:=true
-Bundle-Version: 4.3.1.qualifier
+Bundle-Version: 5.0.0.qualifier
 Bundle-ClassPath: .
 Bundle-Activator: org.osate.analysis.resource.budgets.ResourceBudgetPlugin
 Bundle-Vendor: CMU/SEI
@@ -15,8 +15,7 @@ Require-Bundle: org.eclipse.ui;bundle-version="[3.115.0,4.0.0)",
  org.osate.ui;bundle-version="[6.5.0,7.0.0)",
  org.osate.aadl2.modelsupport;bundle-version="[8.1.0,9.0.0)",
  org.osate.xtext.aadl2.properties;bundle-version="[3.3.0,4.0.0)",
- org.osate.analysis.architecture;bundle-version="[3.1.0,4.0.0)",
- org.osate.analysis.flows;bundle-version="[6.0.0,7.0.0)",
+ org.osate.analysis.reporting;bundle-version="[1.0.0,2.0.0)",
  org.osate.results;bundle-version="[3.2.0,4.0.0)",
  org.osate.aadl2.contrib;bundle-version="[2.2.0,3.0.0)",
  org.osate.contribution.sei;bundle-version="[2.2.0,3.0.0)",
