+++ analyses/org.osate.analysis.resource.budgets/pom.xml
@@ -12,7 +12,7 @@
 
 	<groupId>org.osate</groupId>
 	<artifactId>org.osate.analysis.resource.budgets</artifactId>
-	<version>4.3.1-SNAPSHOT</version>
+	<version>5.0.0-SNAPSHOT</version>
 	<packaging>eclipse-plugin</packaging>
 
 	<build>
