+++ core/org.osate.help/pom.xml
@@ -114,6 +114,11 @@
 		</dependency>
 		<dependency>
 			<groupId>org.osate</groupId>
+			<artifactId>org.osate.analysis.reporting</artifactId>
+			<version>1.0.0-SNAPSHOT</version>
+		</dependency>
+		<dependency>
+			<groupId>org.osate</groupId>
 			<artifactId>org.osate.analysis.flows</artifactId>
 			<version>5.2.0-SNAPSHOT</version>
 		</dependency>
